Update Final Project.py
Added comments and made the layout work. The only thing at this point is to make it so that when the CSV gets updates, an entry is made saying of they want to be on the mailing list or not. For some reason I cant get the buttons to work right for that.